Add .avatar? to Profile and delegate from User
Just to make it simpler to determine if a profile has an avatar by
eliminating one link in the dot chain. So instead of

```
user.profile.avatar.attached?
```

we can now call

```
user.profile.avatar?
```

Then we delegate the call from User to the Profile, leaving us with

```
user.avatar?
```